refactor: collapse flow state machine + on_success into 2 files
Merges flow_state_machine_runtime.go into flow_state_machine.go and
folds flow_on_success_create_user.go, flow_on_success_verify_credentials.go,
and flow_password_hasher.go into flow_on_success.go. No behaviour
changes — same exported surface, same tests pass.
